+++ indico/modules/events/registration/util_test.py
@@ -726,12 +726,13 @@ def test_get_user_data(monkeypatch, dummy_event, dummy_user, dummy_regform):
     dummy_user.title = None
     user_data = get_user_data(dummy_regform, dummy_user, invitation)
     assert user_data == {'email': 'awang@example.test', 'first_name': 'Amy', 'last_name': 'Wang',
-                         'phone': '+1 22 50 14', 'address': 'Geneva', 'affiliation': 'ACME Inc.'}
+                         'phone': '+1 22 50 14', 'address': 'Geneva',
+                         'affiliation': {'id': None, 'text': 'ACME Inc.'}}
 
     # Check that data is taken from the invitation when user is missing
     user_data = get_user_data(dummy_regform, None, invitation)
     assert user_data == {'email': 'awang@example.test', 'first_name': 'Amy', 'last_name': 'Wang',
-                         'affiliation': 'ACME Inc.'}
+                         'affiliation': {'id': None, 'text': 'ACME Inc.'}}
 
     # Check that data from disabled/deleted fields is not used
     title_field = next(item for item in dummy_regform.active_fields
