+++ indico/modules/users/controllers.py
@@ -55,11 +55,11 @@ from indico.modules.users.operations import (add_secondary_email, create_user, d
                                              revoke_admin)
 from indico.modules.users.schemas import (AffiliationArgs, AffiliationSchema, BasicCategorySchema, FavoriteEventSchema,
                                           UserPersonalDataSchema)
-from indico.modules.users.util import (GravatarError, get_avatar_url_from_name, get_gravatar_for_user,
-                                       get_linked_events, get_mastodon_server_name, get_related_categories,
-                                       get_suggested_categories, get_unlisted_events, get_user_by_email,
-                                       get_user_titles, log_user_update, merge_users, search_affiliations, search_users,
-                                       send_avatar, serialize_user, set_user_avatar)
+from indico.modules.users.util import (GravatarError, SearchAffiliationsMixin, get_avatar_url_from_name,
+                                       get_gravatar_for_user, get_linked_events, get_mastodon_server_name,
+                                       get_related_categories, get_suggested_categories, get_unlisted_events,
+                                       get_user_by_email, get_user_titles, log_user_update, merge_users,
+                                       search_affiliations, search_users, send_avatar, serialize_user, set_user_avatar)
 from indico.modules.users.views import (WPAffiliationsDashboard, WPUser, WPUserDashboard, WPUserDataExport,
                                         WPUserFavorites, WPUserPersonalData, WPUserProfilePic, WPUsersAdmin)
 from indico.util.countries import get_countries
@@ -321,20 +321,6 @@ class RHPersonalDataUpdate(RHUserBase):
             log_user_update(self.user, changes)
         flash(_('Your personal data was successfully updated.'), 'success')
         return '', 204
-
-
-class SearchAffiliationsMixin:
-    @use_kwargs({'q': fields.String(load_default='')}, location='query')
-    def _process(self, q):
-        filters = values_from_signal(signals.affiliations.get_affiliation_filters.send(self, context=self.context),
-                                     as_list=True, multi_value_types=list)
-        res = search_affiliations(q, filters=filters)
-        basic_fields = ('id', 'name', 'code', 'street', 'postcode', 'city', 'country_code', 'meta')
-        return AffiliationSchema(many=True, only=basic_fields).jsonify(res)
-
-    @property
-    def context(self):
-        raise NotImplementedError
 
 
 class RHSearchAffiliations(SearchAffiliationsMixin, RH):
